fix(security): Stripe webhook replay protection + private resume storage
Security audit findings, both fixed with regression tests:

1. Stripe webhook idempotency was resulting-status-only (safe against an
   exact re-delivery, but an out-of-order replay of an older, still validly
   signed event -- e.g. a captured payment_intent.canceled resent after the
   escrow was already released -- could silently revert a terminal payment
   status). Adds a stripe_webhook_events dedup table (event_id primary key,
   insert-before-process) plus a guard that refuses to move a payment out of
   a terminal status (RELEASED/REFUNDED/FAILED) via webhook.

2. EngineerService.upload_resume() persisted a permanent, unauthenticated
   public URL (MINIO_PUBLIC_ENDPOINT/bucket/key) as resume_url, and every
   profile response serialized it directly -- if the resumes bucket is
   public-read (required for that link to work at all), anyone who ever
   saw the URL (browser history, referrer header, server log, screenshot)
   could fetch the file forever with no auth and no expiration. Now only
   the private object key is stored; resume_url is always served through a
   freshly generated, 15-minute presigned URL computed after the existing
   owner/admin authorization check, with a compatibility path for rows
   still holding the old permanent-URL format.

Migration 032 adds stripe_webhook_events; ruff/mypy clean; full backend
test suite green (one pre-existing, unrelated failure in
test_work_lifecycle_consolidation.py confirmed present on dev before this
change).

+++ apps/api/app/core/storage.py
@@ -58,7 +58,7 @@ def ensure_bucket_exists(bucket_name: str) -> None:
 def generate_presigned_url(
     bucket_name: str,
     object_name: str,
-    expires_hours: int = 1,
+    expires_hours: float = 1,
 ) -> str | None:
     """Generate a presigned GET URL for temporary file access."""
     client = get_s3_client()
@@ -66,7 +66,7 @@ def generate_presigned_url(
         return client.generate_presigned_url(
             "get_object",
             Params={"Bucket": bucket_name, "Key": object_name},
-            ExpiresIn=expires_hours * 3600,
+            ExpiresIn=int(expires_hours * 3600),
         )
     except ClientError as e:
         logger.error(
